refactor join game to own file
